requirements.txt
accelerate==1.10.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.16
decorator==5.2.1
Deprecated==1.2.18
dill==0.4.0
distlib==0.4.0
distro==1.9.0
docling==2.54.0
docling-core==2.48.2
docling-ibm-models==3.9.1
docling-parse==4.5.0
easyocr==1.7.2
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
executing==2.2.1
Faker==37.8.0
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.31.1
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-groq==0.3.8
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langdetect==1.0.9
langsmith==0.4.29
latex2mathml==3.78.1
lazy_loader==0.4
lxml==5.4.0
Markdown==3.9
markdown-it-py==4.0.0
marko==2.2.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
mpire==2.10.2
mpmath==1.3.0
msoffcrypto-tool==5.4.2
multidict==6.6.4
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==2.5.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
nltk==3.9.1
numpy==2.2.6
olefile==0.47
omegaconf==2.3.0
onnx==1.19.0
onnxruntime==1.22.1
openai==1.108.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.2
parso==0.8.5
pdf2image==1.17.0
pdfminer.six==20250506
pi_heif==1.1.0
pikepdf==9.11.0
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
polyfactory==2.22.2
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycocotools==2.0.10
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pylatexenc==2.10
PyMuPDF==1.26.4
pypandoc==1.15
pyparsing==3.2.4
pypdf==6.1.0
pypdfium2==4.30.0
pyproject-api==1.9.1
pyreadline3==3.5.4
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.1.0
RapidFuzz==3.14.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
rtree==1.4.1
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.2
semchunk==2.2.2
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
stack-data==0.6.3
streamlit==1.49.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tifffile==2025.9.9
tiktoken==0.11.0
timm==1.0.19
tokenizers==0.22.0
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
typer==0.16.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.18.14
unstructured-client==0.42.3
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3==2.5.0
virtualenv==20.34.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.14
wrapt==1.17.3
xlrd==2.0.2
xlsxwriter==3.2.8
yarl==1.20.1
zstandard==0.25.0
